pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "killer-wail"
version = "0.1.0"
description = "Hot-swap orchestrator for Orca-c → Bitwig on macOS"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "watchdog>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
kw = "killer_wail.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
